pyproject.toml
[project]
name = "qwen3-vl-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.128.0",
    "gradio>=6.4.0",
    "gradio-client>=2.0.3",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
    "transformers-stream-generator>=0.0.5",
    "uvicorn>=0.40.0",
]
